added shuffle parameter to records generator
