+++ tests/test_io.py
@@ -1,13 +1,13 @@
 import wfdb
-import numpy as np
 import pytest
-from database_tools.io.records import generate_record_paths, get_data_record, get_header_record, get_signal
+import numpy as np
+from database_tools.io.records import generate_record_paths, get_data_record, get_header_record, header_has_signals, get_signal
 
 NoneType = type(None)
 
 
-@pytest.mark.parametrize('input,expected', [(None, list),
-                                            ('adults', list),
+@pytest.mark.parametrize('input,expected', [(None,       list),
+                                            ('adults',   list),
                                             ('neonates', list)])
 def test_generate_record_paths(input, expected):
     result = [x for x in generate_record_paths(name=input)]
@@ -17,9 +17,17 @@ def test_generate_record_paths(input, expected):
                                             (('30/3000303',              'layout'), NoneType),
                                             (('30/3000063',              'data'),   wfdb.MultiRecord),
                                             (('30/3000063/3000063_0016', 'data'),   wfdb.Record)])
-def test_get_layout_record(input, expected):
+def test_get_header_record(input, expected):
     result = get_header_record(path=input[0], record_type=input[1])
     assert isinstance(result, expected)
+
+@pytest.mark.parametrize('input,expected', [(('30/3000063',               'layout', ['PLETH', 'ABP']),  True),
+                                            (('30/3000063',               'layout', ['PLETH', 'FAIL']), False),
+                                            (('30/3000063/3000063_0016',  'data',   ['PLETH', 'ABP']),  True)])
+def test_header_has_signals(input, expected):
+    hea = get_header_record(path=input[0], record_type=input[1])
+    result = header_has_signals(hea, signals=input[2])
+    assert result == expected
 
 @pytest.mark.parametrize('input,expected', [('30/3000063/3000063_0016', wfdb.Record),
                                             ('30/3000003/3000063_0016', NoneType)])
@@ -28,7 +36,7 @@ def test_get_data_record(input, expected):
     assert isinstance(result, expected)
 
 @pytest.mark.parametrize('input,expected,exception', [(('30/3000063/3000063_0016', 'PLETH'), np.ndarray, None),
-                                                      (('30/3000003/3000063_0016', 'FAIL'),   None,       ValueError)])
+                                                      (('30/3000003/3000063_0016', 'FAIL'),  None,       ValueError)])
 def test_get_signal(input, expected, exception):
     rcd = get_data_record(path=input[0])
     if expected is not None:
